some windows fixes
fixes:

* zipfile was not closed
* path in archive and in fs is different (toto/tata in archive and
toto\tata in FS on windows) so we extract from the zip_path
in the archive and then we move to the path we want. The move is
necessary because zipfile can modify the name in the archive (for
instance special character can be changed into _ on windows) and you
can't set the output path.
* binaries have a different name depending of the platform: windows
binaries end with .exe