use file chooser
(on travis we use temporary thiolliere pyinstaller repo. for macos)

code refactor:
download is no longer needed as we only download one file (it was useful
when we were also downloading vexpress boot)

a file chooser button ask for build directory

an image with extensio building.img is created

extension is changed to error.img or .img at the end of the process.

we do not delete the file even on sd card installation.