impl create admin account Fix #100
a new argument is send to run_installation:
it can be None or a dict with login and pwd

login and pwd are very restricited because I don't really know what is
permitted.

in order not to show the password a new argument is added to the
exec_cmd: show_command.