Force the lang of ideascube.
`LANGUAGE_CODE` is the default language of ideascube if nothing else is
asking for a specific language.

As the `accepted-language` http-header is a ask for a specific language,
it seems that most of the time, the `LANGUAGE_CODE` is useless.

By enabling only the wanted language we will force ideascube to use the
language we want.