Fixed #513: subprocess_pretty_call to not use subprocess.wait
According to doc, subprocess.wait() is not safe under some circumstances ; when the
ouput is large.
This prevent usdisks dump from working for some users on Linux

This uses subprocess.communicate() instead which has the disadvantage of not returning
all output line one by one, but stdout and stderr aside instead.
We're thus return STDERR after stdout.

Also, content is always decoded as the non-decoded version was not used anywhere.
stdin param and handling removed for same reason as well.

Ref: https://docs.python.org/3/library/subprocess.html#subprocess.Popen.wait