Fixed #585: fixed libmount conflict on linux
pyinstaller binary includes a libmount.so.1 as a dependency to some of the
glib shared libraries that gets included with pygobjects.

this file conflicts with the host's libmount that the host's `mount` utility requires
to mount our data parttition. Unless the running host is the same as the building version
(ubuntu-18.04 as of this writting), this results in `mount` not working.

the binary is self-extracted on start then our script is launched with updated
environ so that our app can find included libraries. we have no control over this.

This commit thus deletes the libmount.so.1 file right after extraction (before the main app starts).
There might be unforseen consequences if our app need to use libmount through pygobject but it's unlikely:
we don't do any such thing ourselves and I don't think our limited UI gives any such ability (auto-mount a devices from the folder selection panel maybe??)