Fixed #598: added HTTPS support for all vhosts
All vhost now reachable via HTTPS URLs, using the single, shared certicicate
that we used for the captive portal.

Also updated the list of altnames in the certificate to include all of those but that's
not strictly necessary as our self-signed certificate is already refused by browsers
for that reason and the manual bypass is general.