Fixed #597: Quit button not working
Once the installation completes, whether successfully or not, a new dialog is displayed
offering to start a new run or to Quit.

This quit button, before quitting the app, calls cancel() on the process to make sure
we release everything that needs to.

A race condition or an unhandled termination of qemu might leave it's PID in the lost
of the processes we monitor and that we'd be killing then.

Attempting to kill a non-running PID would crash this cancel task, preventing the actual
exit of the app.

This fixes it by making sure such an exception is captured